added product page
