+++ lib/home/page/home_page.dart
@@ -1,1 +1,26 @@
-// TODO Implement this library.
+import 'package:flutter/material.dart';
+import 'package:reklama_app/add_product/page/add_product_page.dart';
+
+class HomePage extends StatelessWidget {
+  const HomePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('HomePage'),
+      ),
+      floatingActionButton: FloatingActionButton(
+        child: const Icon(Icons.add),
+        onPressed: () {
+          Navigator.push(
+            context,
+            MaterialPageRoute(
+              builder: (context) => AddProductPage(),
+            ),
+          );
+        },
+      ),
+    );
+  }
+}
